Allow passing in convertToRaw function to convertToHTML
Because Entity state is global to the instance of Draft, if a user loads
entities into their Draft and then tries to call convertToHTML on that
content state, our Draft's instance won't have those Entities, and our
convertToRaw will blow up.

So we need to allow users to pass in their own convertToRaw, which will
access their Entity state.