Ensure correct block keys in selectionState after conversion from HTML
When converting HTML to a ContentBlock, we start with an empty block to
get a starting ContentState object. However, the starting block is
overwritten when we set the 'blockMap' based on the blocks created from
the HTML. At this point, the 'selectionBefore' and 'selectionAfter'
properties of the ContentBlock still reference the original block's key
which is no longer valid. This change ensures that we correctly set the
keys in 'selectionBefore' and 'selectionAfter' to match the first block
of the ContentState. There's also a test added to check this case.